// frontend/src/api/openapi/models/FolderUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FolderUpdate = {
  /**
   * Updated name of the folder
   */
  name?: string | null;
  /**
   * Updated parent folder ID
   */
  parent_id?: number | null;
};